[refactor] CalendarService 인터페이스 - 구현체 구조로 변경
